refactorize code

+++ spring-mvc-sample/src/main/java/com/asierg/spring/editor/PricePropertyEditor.java
@@ -1,29 +1,29 @@
 package com.asierg.spring.editor;
 
-import java.beans.PropertyEditorSupport;
-
 import org.springframework.util.StringUtils;
+
+import java.beans.PropertyEditorSupport;
 
 public class PricePropertyEditor extends PropertyEditorSupport {
 
-	public PricePropertyEditor() {
+    public PricePropertyEditor() {
 
-	}
+    }
 
-	public void setAsText(String priceText) {
-		if (!StringUtils.isEmpty(priceText)) {
-			setValue(Double.parseDouble(priceText));
-		} else {
-			setValue(null);
-		}
-	}
+    public void setAsText(String priceText) {
+        if (!StringUtils.isEmpty(priceText)) {
+            setValue(Double.parseDouble(priceText));
+        } else {
+            setValue(null);
+        }
+    }
 
-	public String getAsText() {
-		if ((Double) getValue() != null) {
-			return String.valueOf((Double) getValue());
-		} else {
-			return null;
-		}
-	}
+    public String getAsText() {
+        if (getValue() != null) {
+            return String.valueOf((Double) getValue());
+        } else {
+            return null;
+        }
+    }
 
 }
